refactor: Remove utils-dump
