+++ src/routes/_authenticated/courses/$courseSlug/journal.tsx
@@ -1,4 +1,4 @@
-import { createCourseSectionBreadcrumb } from '@/header/header-data.utils';
+import { createCourseSectionBreadcrumb } from '@/course/course-header';
 import { createFileRoute } from '@tanstack/react-router';
 
 export const Route = createFileRoute(
