+++ src/routes/_authenticated/courses/$courseSlug/repositories/$studentUsername/commits/$commitId.tsx
@@ -1,4 +1,4 @@
-import { getRepositoryCommitDetailBreadcrumb } from '@/header/header-data.utils';
+import { getRepositoryCommitDetailBreadcrumb } from '@/course/course-header';
 import { createFileRoute } from '@tanstack/react-router';
 
 export const Route = createFileRoute(
