+++ src/routes/_authenticated/courses/$courseSlug/tasks/$taskId/edit.tsx
@@ -1,5 +1,5 @@
 import { requireCourseRole } from '@/course/course.guards';
-import { getTaskEditBreadcrumb } from '@/header/header-data.utils';
+import { getTaskEditBreadcrumb } from '@/course/course-header';
 import { createFileRoute } from '@tanstack/react-router';
 
 export const Route = createFileRoute(
