+++ src/routes/_authenticated/index.tsx
@@ -1,5 +1,4 @@
 import { CourseCard, CourseCardSkeleton } from '@/course/course-card.component';
-import { getDashboardBreadcrumb } from '@/header/header-data.utils';
 import { useCoursesQuery } from '@/course/course.queries';
 import { DeadlinesCalendar } from '@/deadlines-calendar/deadlines-calendar.component';
 import { createFileRoute } from '@tanstack/react-router';
@@ -9,7 +8,7 @@ import type { ReactNode } from 'react';
 export const Route = createFileRoute('/_authenticated/')({
   staticData: {
     header: {
-      getBreadcrumb: getDashboardBreadcrumb,
+      getBreadcrumb: [{ label: 'Панель управления' }],
     },
   },
   component: RouteComponent,
